refactor(auth): move login page to /login route
