--- src/app/page.module.css
@@ -1,179 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  height: 100vh;
-  background-color: #f5f5f5;
-  overflow: hidden;
-}
-
-.container_left {
-  padding: 20% 19%;
-  z-index: 1;
-  position: relative;
-  width: 50%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  overflow: hidden;
-}
-
-.container_left > div:first-child {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  z-index: 1;
-}
-
-.container_right {
-  width: 50%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: white;
-  box-shadow: -4px 0 10px rgba(0, 0, 0, 0.1);
-}
-
-.text {
-  position: relative;
-  z-index: 2;
-  text-align: center;
-  color: white;
-  width: 100%;
-  padding: 2rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  transform: translateY(-50%);
-  top: 50%;
-}
-
-.text h1 {
-  font-size: 2.5rem;
-  margin-bottom: 1rem;
-  font-weight: 700;
-}
-
-.text p {
-  font-size: 1.1rem;
-  margin-bottom: 0.5rem;
-}
-
-.register_link {
-  color: #007bff;
-  text-decoration: none;
-  font-weight: 500;
-  transition: color 0.3s ease;
-  display: inline-block;
-  margin-top: 0.5rem;
-}
-
-.register_link:hover {
-  color: #0056b3;
-}
-
-.form {
-  width: 100%;
-  max-width: 400px;
-  padding: 2rem;
-}
-
-.form h2 {
-  font-size: 2rem;
-  margin-bottom: 2rem;
-  color: #333;
-  text-align: center;
-}
-
-.form_group {
-  margin-bottom: 1.5rem;
-}
-
-.form_group label {
-  display: block;
-  margin-bottom: 0.5rem;
-  color: #666;
-  font-weight: 500;
-}
-
-.form_group input {
-  width: 94%;
-  padding: 0.75rem;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  font-size: 1rem;
-  transition: border-color 0.3s ease;
-}
-
-.form_group input:focus {
-  outline: none;
-  border-color: #007bff;
-  box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.25);
-}
-
-.submit_button {
-  width: 100%;
-  padding: 0.75rem;
-  background-color: #007bff;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  font-size: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.submit_button:hover {
-  background-color: #0056b3;
-}
-
-.submit_button:disabled {
-  background-color: #ccc;
-  cursor: not-allowed;
-}
-
-.error {
-  background-color: #fff3f3;
-  color: #dc3545;
-  padding: 0.75rem;
-  border-radius: 4px;
-  margin-bottom: 1rem;
-  border: 1px solid #dc3545;
-  text-align: center;
-}
-
-.text{
-  color: black;
-}
-
-.password_input_container {
-  position: relative;
-  display: flex;
-  align-items: center;
-}
-
-.password_input_container input {
-  width: 100%;
-  padding-right: 2.5rem;
-}
-
-.password_toggle {
-  position: absolute;
-  right: 10px;
-  background: none;
-  border: none;
-  cursor: pointer;
-  padding: 5px;
-  font-size: 1.2rem;
-  color: #666;
-  transition: color 0.3s ease;
-}
-
-.password_toggle:hover {
-  color: #007bff;
-}
